[jUnit] restructuring and add NewsDao
