+++ src/main/java/com/epam/jmp/concurrency/job/Import.java
@@ -3,8 +3,8 @@ package com.epam.jmp.concurrency.job;
 import static com.epam.jmp.util.Config.INPUT_FOLDER;
 
 import com.epam.jmp.concurrency.files.FolderReader;
-import com.epam.jmp.concurrency.service.LocalNewsService;
-import com.epam.jmp.concurrency.service.NewsService;
+import com.epam.jmp.concurrency.model.service.LocalNewsService;
+import com.epam.jmp.concurrency.model.service.NewsService;
 import com.epam.jmp.util.Config;
 
 public class Import implements Runnable {
