+++ src/main/java/com/epam/jmp/concurrency/model/dao/NewsDao.java
@@ -1,0 +1,9 @@
+package com.epam.jmp.concurrency.model.dao;
+
+import com.epam.jmp.concurrency.model.News;
+
+public interface NewsDao {
+
+	News mergeNews(News news);
+
+}
